Create objects folder with objects classes
